refactor: move PersonType para um local mais adequado.
